refactor: update the code dependent on light client
