+++ internal/statesync/stateprovider.go
@@ -13,15 +13,15 @@ import (
 	dbm "github.com/tendermint/tm-db"
 
 	dashcore "github.com/tendermint/tendermint/dash/core"
+	"github.com/tendermint/tendermint/dash/light"
+	lightprovider "github.com/tendermint/tendermint/dash/light/provider"
+	lighthttp "github.com/tendermint/tendermint/dash/light/provider/http"
+	lightrpc "github.com/tendermint/tendermint/dash/light/rpc"
+	lightdb "github.com/tendermint/tendermint/dash/light/store/db"
 	"github.com/tendermint/tendermint/internal/p2p"
 	sm "github.com/tendermint/tendermint/internal/state"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/libs/log"
-	"github.com/tendermint/tendermint/light"
-	lightprovider "github.com/tendermint/tendermint/light/provider"
-	lighthttp "github.com/tendermint/tendermint/light/provider/http"
-	lightrpc "github.com/tendermint/tendermint/light/rpc"
-	lightdb "github.com/tendermint/tendermint/light/store/db"
 	ssproto "github.com/tendermint/tendermint/proto/tendermint/statesync"
 	rpchttp "github.com/tendermint/tendermint/rpc/client/http"
 	"github.com/tendermint/tendermint/types"
